Airflow/requirements.txt
pandas
apache-airflow-providers-mongo
apache-airflow-providers-amazon
pycaret
pycaret[mlops]